# contracts/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Self-hosted SRE/observability aggregator core API.",
    "title": "sre-kit API",
    "contact": {},
    "version": "1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "projects"
        ],
        "summary": "List projects",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/internal_projects_interfaces_http.response"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "projects"
        ],
        "summary": "Create project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/internal_projects_interfaces_http.request"
              }
            }
          },
          "description": "project",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_projects_interfaces_http.response"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "delete": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "projects"
        ],
        "summary": "Delete project",
        "parameters": [
          {
            "description": "project ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "projects"
        ],
        "summary": "Update project",
        "parameters": [
          {
            "description": "project ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/internal_projects_interfaces_http.request"
              }
            }
          },
          "description": "project fields",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_projects_interfaces_http.response"
                }
              }
            }
          }
        }
      }
    },
    "/api/sources": {
      "get": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "List every configured source and its current status",
        "tags": [
          "sources"
        ],
        "summary": "List sources",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/internal_sources_interfaces_http.sourceResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Configure a new adapter instance",
        "tags": [
          "sources"
        ],
        "summary": "Create a source",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/internal_sources_interfaces_http.createSourceRequest"
              }
            }
          },
          "description": "adapter_id + config",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_sources_interfaces_http.sourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/{id}": {
      "delete": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Remove a configured source permanently",
        "tags": [
          "sources"
        ],
        "summary": "Delete a source",
        "parameters": [
          {
            "description": "source ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Enable/disable a source and/or update its config",
        "tags": [
          "sources"
        ],
        "summary": "Update a source",
        "parameters": [
          {
            "description": "source ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/internal_sources_interfaces_http.updateSourceRequest"
              }
            }
          },
          "description": "fields to patch",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_sources_interfaces_http.sourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "name": "session",
        "in": "cookie"
      }
    },
    "schemas": {
...
      "internal_projects_interfaces_http.request": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "internal_projects_interfaces_http.response": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "internal_sources_interfaces_http.createSourceRequest": {
        "type": "object",
        "properties": {
          "adapter_id": {
            "type": "string"
          },
          "config": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "project_id": {
            "type": "string"
          }
        }
      },
      "internal_sources_interfaces_http.sourceResponse": {
        "type": "object",
        "properties": {
          "adapter_id": {
            "type": "string"
          },
          "config": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "last_seen_at": {
            "type": "string"
          },
          "last_status": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          }
        }
      },
      "internal_sources_interfaces_http.updateSourceRequest": {
        "type": "object",
        "properties": {
          "config": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}